perf: optimize input

+++ Assets/_IM/Scripts/Input/LocalInputProvider.cs
@@ -8,10 +8,10 @@ namespace StinkySteak.IM.PlayerInput
         public override void OnInput(NetworkSandbox sandbox)
         {
             PlayerCharacterInput input = new();
-            input.Right = Input.GetAxisRaw("Horizontal");
-            input.Forward = Input.GetAxisRaw("Vertical");
-            input.Up = Input.GetAxisRaw("Flying");
-            input.Yaw = Input.GetAxisRaw("Yaw");
+            input.Right = (int)Input.GetAxisRaw("Horizontal");
+            input.Forward = (int)Input.GetAxisRaw("Vertical");
+            input.Up = (int)Input.GetAxisRaw("Flying");
+            input.Yaw = (int)Input.GetAxisRaw("Yaw");
 
             sandbox.SetInput(input);
         }
